requirements.txt
imageio==2.9.0
tqdm==4.50.2
scipy==1.5.2
numpy==1.19.2
matplotlib==3.4.1
neptune_client==0.9.11
easydict==1.9
GPUtil==1.4.0
Pillow==8.3.1
PyYAML==5.4.1
scikit_learn==0.24.2
torch==1.11.0
tensorboard==2.6.0
tables==3.7.0
pandas==1.4.2
h5py==3.6.0